requirements.txt
langchain
openai
tiktoken
chromadb
pypdf
pyautogen
huggingface_hub
sentence-transformers
transformers
google-generativeai
einops
ctransformers
streamlit
sentencepiece
datasets[audio]
torch
torchvision
torchaudio
openai
requests
pillow
pydub
stqdm
diffusers
